move to own file and create tests for Vehicle
