+++ app/core/models/old_models.py
@@ -1,25 +1,8 @@
 from django.db.models import *
 
-from core.models.company import Company
-from core.models.leasing_mode import LeasingMode
-from core.models.plan import Plan
 from core.models.route import Route
 from core.models.station import Station
 from core.models.tender import Tender
-from core.models.vehicle_type import VehicleType
-
-
-class Vehicle(Model):
-    """The Vehicle class is for "instances" of vehicles. Every instance has a VehicleClass, a leasing_mode and an
-    owner. """
-    plan = ForeignKey(Plan, null=True, on_delete=SET_NULL)
-    vtype = ForeignKey(VehicleType, on_delete=PROTECT)
-    owner = ForeignKey(Company, on_delete=PROTECT)
-    leasing_mode = ForeignKey(LeasingMode, on_delete=PROTECT)
-    leased_since = DateTimeField()
-
-    def __str__(self):
-        return self.vtype.name + "-" + self.id
 
 
 class Criterion(Model):
